requirements.txt
# Web & API
fastapi
uvicorn[standard]
python-multipart  # Para recibir archivos PDF
python-jose[cryptography] # JWT para Auth
passlib[bcrypt]    # Hash de contraseñas
python-dotenv

# Base de Datos
sqlmodel          # SQLAlchemy + Pydantic (perfecto para FastAPI)
psycopg2-binary   # Driver para PostgreSQL
alembic           # Migraciones

# Procesamiento y LLM
polars
pymupdf
pdfplumber
langchain
langchain-openai
langchain-groq
pydantic

# Visualización y Testeo
plotly
pytest
httpx
weasyprint
structlog
asyncpg
